langchain-translator/requirements.txt
langchain
langchain-openai
langchain_core
python-dotenv
ipykernel
langserve[all]
fastapi